pyproject.toml
[project]
name = "chess-experiments"
version = "0.1.0"
description = "Chess AI experiments: probing, evaluation metrics, and interpretability."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = ""},
]
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.8.2",
    "matplotlib>=3.9.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=6.29.5",
    "lczerolens",
    "loguru>=0.7.3",
    "scikit-learn>=1.8.0",
    "tdhook",
    "wandb>=0.19.9",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]

[build]
target-dir = "build/dist"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
default-groups = ["dev"]

[tool.uv.sources]
tdhook = { git = "https://github.com/Xmaster6y/tdhook" }
lczerolens = { git = "https://github.com/Xmaster6y/lczerolens" }

[tool.ruff]
line-length = 119
target-version = "py311"
